fix(e2e): Patch OLM Subscription to avoid conflicting updates
